Modify : api 테스트 및 UI 수정
