+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^12.1.10",
     "font-awesome": "^4.7.0",
     "node-sass": "^4.14.1",
+    "pure-react-carousel": "^1.27.6",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-icons": "^4.1.0",
