--- public/data/bookList.json
@@ -1,64 +1,0 @@
-{
-  "BOOKS" : [
-      {
-        "id": 1,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1605809484/gb2tzhihpognmcytsar4.jpg",
-        "title": "셀트리오니즘",
-        "rating": "2.5"
-      },
-      {
-        "id": 2,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1607306887/iiwp0mpsalrcydkqujew.jpg",
-        "title": "일인칭 단수",
-        "rating": "3.5"
-      },
-      {
-        "id": 3,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1592041960/jghkbab0mlbqaphhyoqt.jpg",
-        "title": "공정하다는 착각",
-        "rating": "2.5"
-      },
-      {
-        "id": 4,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_700,q_80,w_490/v1466062869/b849qbxyu6iq2ecsymah.jpg",
-        "title": "달러구트 꿈 백화점",
-        "rating": "2.5"
-      },
-      {
-        "id": 5,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1592041960/jghkbab0mlbqaphhyoqt.jpg",
-        "title": "미드나잇 선",
-        "rating": "4.2"
-      },
-      {
-        "id": 6,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_700,q_80,w_490/v1466062869/b849qbxyu6iq2ecsymah.jpg",
-        "title": "미드나잇 선",
-        "rating": "4.2"
-      },
-      {
-        "id": 7,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_700,q_80,w_490/v1466062869/b849qbxyu6iq2ecsymah.jpg",
-        "title": "미드나잇 선",
-        "rating": "4.2"
-      },
-      {
-        "id": 8,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1596856934/uu8edn7drm79ehbpiexz.jpg",
-        "title": "미드나잇 선",
-        "rating": "4.2"
-      },
-      {
-        "id": 9,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1598941919/aoyd1bild4aedugvqeky.jpg",
-        "title": "미드나잇 선",
-        "rating": "4.2"
-      },
-      {
-        "id": 10,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1502242918/ity96bqakynpczzd3p2p.jpg",
-        "title": "미드나잇 선",
-        "rating": "4.2"
-      }
-    ]
-}
