+++ public/data/movieList.json
@@ -2,7 +2,7 @@
   "MOVIES" : [
       {
         "id": 1,
-        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466148147/me7i5jkb3i1uolm3mwun.jpg",
+        "moviePoster": "https://t1.daumcdn.net/cfile/2646D04D52C520292C",
         "title": "위플래쉬",
         "year": "2020",
         "rating": "2.5",
@@ -74,6 +74,86 @@
       },
       {
         "id": 10,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1605408533/lhdypyrxrz8axmdkyyfb.jpg",
+        "title": "도라에몽",
+        "year": "2020",
+        "rating": "2.5",
+        "country": "미국"
+      },
+      {
+        "id": 11,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466148147/me7i5jkb3i1uolm3mwun.jpg",
+        "title": "위플래쉬",
+        "year": "2020",
+        "rating": "2.5",
+        "country": "미국"
+      },
+      {
+        "id": 12,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466148147/me7i5jkb3i1uolm3mwun.jpg",
+        "title": "조제",
+        "year": "2020",
+        "rating": "4.0",
+        "country": "미국"
+      },
+      {
+        "id": 13,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466406561/elfhogfydfrnxhhv3gwz.jpg",
+        "title": "원더 우먼",
+        "year": "2020",
+        "rating": "3.5",
+        "country": "미국"
+      },
+      {
+        "id": 14,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466064258/x8yowmjhm0jnl2ukqans.jpg",
+        "title": "가족의 탄생",
+        "year": "2020",
+        "rating": "2.5",
+        "country": "미국"
+      },
+      {
+        "id": 15,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1484792943/jooyj0i3atc617rh0dte.jpg",
+        "title": "멋진 하루",
+        "year": "2020",
+        "rating": "3.0",
+        "country": "미국"
+      },
+      {
+        "id": 16,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466148147/me7i5jkb3i1uolm3mwun.jpg",
+        "title": "도라에몽",
+        "year": "2020",
+        "rating": "2.5",
+        "country": "미국"
+      },
+      {
+        "id": 17,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1484792943/jooyj0i3atc617rh0dte.jpg",
+        "title": "멋진 하루",
+        "year": "2020",
+        "rating": "3.0",
+        "country": "미국"
+      },
+      {
+        "id": 18,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466148147/me7i5jkb3i1uolm3mwun.jpg",
+        "title": "도라에몽",
+        "year": "2020",
+        "rating": "2.5",
+        "country": "미국"
+      },
+      {
+        "id": 19,
+        "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1466148147/me7i5jkb3i1uolm3mwun.jpg",
+        "title": "멋진 하루",
+        "year": "2020",
+        "rating": "3.0",
+        "country": "미국"
+      },
+      {
+        "id": 20,
         "moviePoster": "https://an2-img.amz.wtchn.net/image/v1/watcha/image/upload/c_fill,h_400,q_80,w_280/v1605408533/lhdypyrxrz8axmdkyyfb.jpg",
         "title": "도라에몽",
         "year": "2020",
