+++ public/index.html
@@ -10,6 +10,9 @@
       content="Web site created using create-react-app"
     />
     <link rel="stylesheet" href="../src/styles/reset.scss"/>
+    <link rel="stylesheet" type="text/css" charset="UTF-8" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css" /> 
+<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css" />
+    
     <title>Youngcha</title>
   </head>
