+++ src/pages/MainPage/Compontents/MoviePages/MovieContainer/MovieCard.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/common.scss';
+@import '../../../../../styles/common.scss';
 
 li {
   display: inline-block;
@@ -8,13 +8,11 @@ li {
 
   .moviePoster {
     margin-bottom: 10px;
-    max-width: 300px;
     overflow: hidden;
 
     img {
-      max-width: 100%;
-      max-height: 100%;
-      height: auto;
+      width: 100%;
+      height: 100%;
       border-radius: 5px;
     }
   }
