+++ src/pages/MainPage/Compontents/MoviePages/MovieContainerBottom/MovieContainerBottom.js
@@ -1,29 +1,6 @@
 import React, { Component } from 'react';
 import MovieCard from '../MovieContainer/MovieCard';
 import Slider from 'react-slick';
-import '../MovieContainer/MovieContainer.scss';
-
-function NextArrow(props) {
-  const { className, style, onClick } = props;
-  return (
-    <div
-      className={className}
-      style={{ ...style, display: 'block' }}
-      onClick={onClick}
-    ></div>
-  );
-}
-
-function PrevArrow(props) {
-  const { className, style, onClick } = props;
-  return (
-    <div
-      className={className}
-      style={{ ...style, display: 'block' }}
-      onClick={onClick}
-    />
-  );
-}
 
 class MovieContainerBottom extends Component {
   render() {
@@ -34,23 +11,29 @@ class MovieContainerBottom extends Component {
       speed: 500,
       slidesToShow: 5,
       slidesToScroll: 5,
-      nextArrow: <NextArrow />,
-      prevArrow: <PrevArrow />,
       responsive: [
         {
           breakpoint: 1024,
           settings: {
+            slidesToShow: 4,
+            slidesToScroll: 4,
+            infinite: true,
+          },
+        },
+        {
+          breakpoint: 800,
+          settings: {
             slidesToShow: 3,
             slidesToScroll: 3,
-            infinite: true,
+            initialSlide: 3,
           },
         },
         {
           breakpoint: 600,
           settings: {
-            slidesToShow: 2,
-            slidesToScroll: 2,
-            initialSlide: 2,
+            slidesToShow: 3,
+            slidesToScroll: 3,
+            initialSlide: 3,
           },
         },
         {
