+++ src/pages/MainPage/Compontents/MoviePages/MoviePage.scss
