+++ src/pages/MainPage/Compontents/TvPages/TvPages.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class TvPages extends Component {
+  render() {
+    return <div></div>;
+  }
+}
+
+export default TvPages;
