+++ src/pages/MainPage/MainPage.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import MoviePage from './components/MainPages/MoviePage';
-import TvProgramPage from './components/MainPages/TvProgramPage';
-import BookPage from './components/MainPages/BookPage';
+import MoviePage from './Compontents/MoviePages/MoviePage';
 import Footer from '../../Components/Footer/Footer';
+import 'slick-carousel/slick/slick.css';
+import 'slick-carousel/slick/slick-theme.css';
 
 class MainPage extends Component {
   render() {
