--- src/pages/MainPage/components/MovieContainer/MovieContainer.scss
@@ -1,29 +1,0 @@
-@import '../../../../styles/common.scss';
-
-/* slide css 설정이라 nesting하지 않았습니다 */
-
-.slick-list {
-  padding: 10px;
-}
-
-.slick-arrow {
-  position: absolute;
-  margin-left: 10px;
-  top: 50%;
-  transform: translateY(-50%);
-  z-index: 1;
-}
-
-.slick-prev:before {
-  @include slide-arrow;
-  content: '<';
-}
-
-.slick-next:before {
-  content: '>';
-  @include slide-arrow;
-}
-
-.slick-next {
-  margin-right: 10px;
-}
